@@ -6,20 +6,11 @@ import (
 	"github.com/watermint/toolbox/domain/dropboxsign/api/hs_conn"
 	"github.com/watermint/toolbox/essentials/api/api_auth"
 	"github.com/watermint/toolbox/essentials/api/api_auth_basic"
-	"github.com/watermint/toolbox/essentials/api/api_conn"
+	"github.com/watermint/toolbox/essentials/api/api_auth_key"
 	"github.com/watermint/toolbox/essentials/api/api_conn_impl"
 	"github.com/watermint/toolbox/essentials/log/esl"
 	"github.com/watermint/toolbox/infra/control/app_control"
 	"github.com/watermint/toolbox/infra/control/app_definitions"
-	"github.com/watermint/toolbox/infra/ui/app_msg"
-)
-
-type MsgDropboxSign struct {
-	AskApiKey app_msg.Message
-}
-
-var (
-	MHelloSign = app_msg.Apply(&MsgDropboxSign{}).(*MsgDropboxSign)
 )
 
 func NewConnHelloSign(name string) hs_conn.ConnHelloSignApi {
@@ -41,7 +32,7 @@ func (z *connHelloSignApi) Connect(ctl app_control.Control) (err error) {
 	l := ctl.Log()
 	sessionData := api_auth.BasicSessionData{
 		AppData: api_auth.BasicAppData{
-			AppKeyName:      api_conn.ServiceDropboxSign,
+			AppKeyName:      app_definitions.ScopeLabelDropboxSign,
 			DontUseUsername: false,
 			DontUsePassword: true,
 		},
@@ -50,7 +41,7 @@ func (z *connHelloSignApi) Connect(ctl app_control.Control) (err error) {
 	entity, mock, err := api_conn_impl.BasicConnect(
 		sessionData,
 		ctl,
-		api_auth_basic.CustomAskUserName(MHelloSign.AskApiKey),
+		api_auth_basic.CustomAskUserName(api_auth_key.MConsole.AskKey),
 	)
 	if mock {
 		z.client = hs_client_impl.NewMock(z.peerName, ctl)
@@ -73,11 +64,11 @@ func (z *connHelloSignApi) SetPeerName(name string) {
 }
 
 func (z *connHelloSignApi) ScopeLabel() string {
-	return app_definitions.ServiceDropboxSign
+	return app_definitions.ScopeLabelDropboxSign
 }
 
-func (z *connHelloSignApi) ServiceName() string {
-	return api_conn.ServiceDropboxSign
+func (z *connHelloSignApi) AppKeyName() string {
+	return app_definitions.AppKeyDropboxSign
 }
 
 func (z *connHelloSignApi) Client() hs_client.Client {
