@@ -1,12 +1,12 @@
 package gh_conn_impl
 
 import (
-	"github.com/watermint/toolbox/domain/github/api/gh_auth"
 	"github.com/watermint/toolbox/domain/github/api/gh_client"
 	"github.com/watermint/toolbox/domain/github/api/gh_client_impl"
 	"github.com/watermint/toolbox/essentials/api/api_auth"
 	"github.com/watermint/toolbox/essentials/api/api_conn"
 	"github.com/watermint/toolbox/infra/control/app_control"
+	"github.com/watermint/toolbox/infra/control/app_definitions"
 )
 
 func NewConnGithubPublic(name string) api_conn.Connection {
@@ -20,12 +20,12 @@ type ConnGithubPublic struct {
 	ctl  app_control.Control
 }
 
-func (z *ConnGithubPublic) ServiceName() string {
-	return api_conn.ServiceGithub
+func (z *ConnGithubPublic) AppKeyName() string {
+	return app_definitions.AppKeyGithubPublic
 }
 
 func (z *ConnGithubPublic) ScopeLabel() string {
-	return gh_auth.ScopeLabelPublic
+	return app_definitions.ScopeLabelGithub
 }
 
 func (z *ConnGithubPublic) PeerName() string {
