@@ -1,32 +1,57 @@
 package app_definitions
 
+// App service keys
+// These keys are used for identifying services in the app key repository.
 const (
-	ServiceDropboxIndividual = "dropbox_individual"
-	ServiceDropboxTeam       = "dropbox_team"
-	ServiceDropboxSign       = "dropbox_sign"
-	ServiceGithub            = "github"
-	ServiceGoogleMail        = "google_mail2024"     // Adding 2024 to avoid conflict with existing auth data #777
-	ServiceGoogleSheets      = "google_sheets2024"   // Adding 2024 to avoid conflict with existing auth data #777
-	ServiceGoogleCalendar    = "google_calendar2024" // Adding 2024 to avoid conflict with existing auth data #777
-	ServiceGoogleTranslate   = "google_translate"
-	ServiceAsana             = "asana"
-	ServiceDeepl             = "deepl"
-	ServiceSlack             = "slack"
-	ServiceFigma             = "figma"
+	AppKeyAsana             = "asana"
+	AppKeyDeepl             = "deepl"
+	AppKeyDropboxIndividual = "dropbox_individual"
+	AppKeyDropboxSign       = "dropbox_sign"
+	AppKeyDropboxTeam       = "dropbox_team"
+	AppKeyFigma             = "figma"
+	AppKeyGithubPublic      = "github_public"
+	AppKeyGithubRepo        = "github_repo"
+	AppKeySlack             = "slack"
 )
 
 var (
-	AllServices = []string{
-		ServiceDropboxIndividual,
-		ServiceDropboxTeam,
-		ServiceDropboxSign,
-		ServiceAsana,
-		ServiceDeepl,
-		ServiceFigma,
-		ServiceGithub,
-		ServiceGoogleCalendar,
-		ServiceGoogleMail,
-		ServiceGoogleSheets,
-		ServiceSlack,
+	AllAppKeys = []string{
+		AppKeyAsana,
+		AppKeyDeepl,
+		AppKeyDropboxIndividual,
+		AppKeyDropboxSign,
+		AppKeyDropboxTeam,
+		AppKeyFigma,
+		AppKeyGithubPublic,
+		AppKeyGithubRepo,
+		AppKeySlack,
+	}
+)
+
+// Scope labels for services
+// These labels are used for categorizing services in the UI, documentation, etc.
+const (
+	ScopeLabelAsana             = "asana"
+	ScopeLabelDeepl             = "deepl"
+	ScopeLabelDropboxIndividual = "dropbox_individual"
+	ScopeLabelDropboxSign       = "dropbox_sign"
+	ScopeLabelDropboxTeam       = "dropbox_team"
+	ScopeLabelFigma             = "figma"
+	ScopeLabelGithub            = "github"
+	ScopeLabelSlack             = "slack"
+	ScopeLabelUtility           = ""
+)
+
+var (
+	Services = []string{
+		ScopeLabelAsana,
+		ScopeLabelDeepl,
+		ScopeLabelDropboxIndividual,
+		ScopeLabelDropboxSign,
+		ScopeLabelDropboxTeam,
+		ScopeLabelFigma,
+		ScopeLabelGithub,
+		ScopeLabelSlack,
+		ScopeLabelUtility,
 	}
 )
