@@ -16,8 +16,6 @@ import (
 	"github.com/watermint/toolbox/domain/figma/api/fg_conn_impl"
 	"github.com/watermint/toolbox/domain/github/api/gh_conn"
 	"github.com/watermint/toolbox/domain/github/api/gh_conn_impl"
-	"github.com/watermint/toolbox/domain/google/api/goog_conn"
-	"github.com/watermint/toolbox/domain/google/api/goog_conn_impl"
 	"github.com/watermint/toolbox/domain/slack/api/work_conn"
 	"github.com/watermint/toolbox/domain/slack/api/work_conn_impl"
 	"github.com/watermint/toolbox/essentials/api/api_conn"
@@ -63,43 +61,31 @@ var (
 		newValueRpModelTransactionReport(""),
 		newValueSelectString(),
 		newValueString(),
-		newValueConn((*as_conn.ConnAsanaApi)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*as_conn.ConnAsanaApi)(nil), func(peerName string) api_conn.Connection {
 			return as_conn_impl.NewConnAsana(peerName)
 		}),
-		newValueConn((*dbx_conn.ConnScopedIndividual)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*dbx_conn.ConnScopedIndividual)(nil), func(peerName string) api_conn.Connection {
 			return dbx_conn_impl.NewConnScopedIndividual(peerName)
 		}),
-		newValueConn((*dbx_conn.ConnScopedTeam)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*dbx_conn.ConnScopedTeam)(nil), func(peerName string) api_conn.Connection {
 			return dbx_conn_impl.NewConnScopedTeam(peerName)
 		}),
-		newValueConn((*deepl_conn.ConnDeeplApi)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*deepl_conn.ConnDeeplApi)(nil), func(peerName string) api_conn.Connection {
 			return deepl_conn_impl.NewConnDeeplApi(peerName)
 		}),
-		newValueConn((*fg_conn.ConnFigmaApi)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*fg_conn.ConnFigmaApi)(nil), func(peerName string) api_conn.Connection {
 			return fg_conn_impl.NewConnFigma(peerName)
 		}),
-		newValueConn((*gh_conn.ConnGithubPublic)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*gh_conn.ConnGithubPublic)(nil), func(peerName string) api_conn.Connection {
 			return gh_conn_impl.NewConnGithubPublic(peerName)
 		}),
-		newValueConn((*gh_conn.ConnGithubRepo)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*gh_conn.ConnGithubRepo)(nil), func(peerName string) api_conn.Connection {
 			return gh_conn_impl.NewConnGithubRepo(peerName)
 		}),
-		newValueConn((*goog_conn.ConnGoogleCalendar)(nil), func(peerName string) api_conn.Connection {
-			return goog_conn_impl.NewConnGoogleCalendar(peerName)
-		}),
-		newValueConn((*goog_conn.ConnGoogleMail)(nil), func(peerName string) api_conn.Connection {
-			return goog_conn_impl.NewConnGoogleMail(peerName)
-		}),
-		newValueConn((*goog_conn.ConnGoogleSheets)(nil), func(peerName string) api_conn.Connection {
-			return goog_conn_impl.NewConnGoogleSheets(peerName)
-		}),
-		newValueConn((*goog_conn.ConnGoogleTranslate)(nil), func(peerName string) api_conn.Connection {
-			return goog_conn_impl.NewConnGoogleTranslate(peerName)
-		}),
-		newValueConn((*hs_conn.ConnHelloSignApi)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*hs_conn.ConnHelloSignApi)(nil), func(peerName string) api_conn.Connection {
 			return hs_conn_impl.NewConnHelloSign(peerName)
 		}),
-		newValueConn((*work_conn.ConnSlackApi)(nil), func(peerName string) api_conn.Connection {
+		NewValueConn((*work_conn.ConnSlackApi)(nil), func(peerName string) api_conn.Connection {
 			return work_conn_impl.NewSlackApi(peerName)
 		}),
 	}
